+++ src/zigbee/examples/zigbee-nwk-routing.cc
@@ -110,8 +110,7 @@ NwkDataIndication(Ptr<ZigbeeStack> stack, NldeDataIndicationParams params, Ptr<P
 static void
 NwkNetworkFormationConfirm(Ptr<ZigbeeStack> stack, NlmeNetworkFormationConfirmParams params)
 {
-    std::cout << "NlmeNetworkFormationConfirmStatus = " << static_cast<uint32_t>(params.m_status)
-              << "\n";
+    std::cout << "NlmeNetworkFormationConfirmStatus = " << params.m_status << "\n";
 }
 
 static void
@@ -152,8 +151,7 @@ NwkNetworkDiscoveryConfirm(Ptr<ZigbeeStack> stack, NlmeNetworkDiscoveryConfirmPa
     }
     else
     {
-        NS_ABORT_MSG(
-            "Unable to discover networks | status: " << static_cast<uint32_t>(params.m_status));
+        NS_ABORT_MSG("Unable to discover networks | status: " << params.m_status);
     }
 }
 
@@ -185,8 +183,7 @@ NwkJoinConfirm(Ptr<ZigbeeStack> stack, NlmeJoinConfirmParams params)
 static void
 NwkRouteDiscoveryConfirm(Ptr<ZigbeeStack> stack, NlmeRouteDiscoveryConfirmParams params)
 {
-    std::cout << "NlmeRouteDiscoveryConfirmStatus = " << static_cast<uint32_t>(params.m_status)
-              << "\n";
+    std::cout << "NlmeRouteDiscoveryConfirmStatus = " << params.m_status << "\n";
 }
 
 int
